+++ apps/ex_marketer_web/assets/js/initializers/index.js
@@ -1,0 +1,1 @@
+import "./socket";
