+++ apps/ex_marketer_web/assets/package.json
@@ -8,7 +8,8 @@
   },
   "dependencies": {
     "phoenix": "file:../../../deps/phoenix",
-    "phoenix_html": "file:../../../deps/phoenix_html"
+    "phoenix_html": "file:../../../deps/phoenix_html",
+    "bootstrap.native": "^3.0.10"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0",
