+++ apps/ex_marketer_web/lib/ex_marketer_web/templates/layout/app.html.eex
@@ -21,6 +21,9 @@
 				<p class="alert alert-danger my-3" role="alert"><%= get_flash(@conn, :error) %></p>
 			<% end %>
 
+			<div class="notification">
+			</div>
+
 			<%= @inner_content %>
 		</main>
 	</body>
